mysql + sqlite parallel

+++ scripts/test-migrations.sh
@@ -31,11 +31,14 @@ function run_tests_for_version() {
     echo "===== Finished testing version $VERSION ====="
 }
 
-# run a mysql instance in docker
-docker run --name mysql -d -p 3306:3306 -e MYSQL_ROOT_PASSWORD=password mysql:latest
+if [ "$1" == "mysql" ]; then
+    echo "===== Testing MySQL ====="
+    # run a mysql instance in docker
+    docker run --name mysql -d -p 3306:3306 -e MYSQL_ROOT_PASSWORD=password mysql:latest
 
-# mysql takes a while to start up
-sleep 30
+    # mysql takes a while to start up
+    sleep 30
+fi
 
 # List of versions to test
 VERSIONS=("0.40.0" "0.40.3" "0.41.0" "0.43.0" "0.44.1" "0.44.3" "0.45.2" "0.45.3" "0.45.4" "0.45.5" "0.45.6" "0.46.0" "0.47.0")
@@ -64,12 +67,17 @@ do
         pip3 install importlib_metadata
     fi
 
-    zenml connect --url mysql://127.0.0.1/zenml --username root --password password
+    if [ "$1" == "mysql" ]; then
+        zenml connect --url mysql://127.0.0.1/zenml --username root --password password
+    fi
+
     # Run the tests for this version
     run_tests_for_version $VERSION
 
-    zenml disconnect
-    sleep 5
+    if [ "$1" == "mysql" ]; then
+        zenml disconnect
+        sleep 5
+    fi
 
     deactivate
 done
@@ -84,9 +92,16 @@ pip3 install -U pip setuptools wheel
 pip3 install -e ".[templates,server]"
 pip3 install importlib_metadata
 
-zenml connect --url mysql://127.0.0.1/zenml --username root --password password
+if [ "$1" == "mysql" ]; then
+    zenml connect --url mysql://127.0.0.1/zenml --username root --password password
+fi
+
 run_tests_for_version current_branch_mysql
+
+if [ "$1" == "mysql" ]; then
+    zenml disconnect
+    docker rm -f mysql
+fi
 
 deactivate
 
-docker rm -f mysql
